remove RepositoryInterface.php and DecoratorInterface.php, replace with EpicInterface - add AbstractDecorator.php
