--- src/Interfaces/DecoratorInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Ulex\EpicRepositories\Interfaces;
-
-interface DecoratorInterface extends EpicInterface
-{
-
-}
