--- src/Interfaces/RepositoryInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Ulex\EpicRepositories\Interfaces;
-
-interface RepositoryInterface extends EpicInterface
-{
-
-}
